requirements.txt
# ─────────────────────────────────────────────
# Core Async & Utilities
# ─────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pyyaml==6.0.2
loguru==0.7.2
rich==13.7.1
httpx==0.27.0
tenacity==8.2.3

# ─────────────────────────────────────────────
# LLM Providers
# ─────────────────────────────────────────────
openai==1.40.3
anthropic==0.30.1
ollama==0.3.0
tiktoken==0.7.0

# ─────────────────────────────────────────────
# Memory / Vector DB
# ─────────────────────────────────────────────
chromadb==0.5.5
sentence-transformers==3.0.1
aiosqlite==0.20.0
numpy==1.26.4

# ─────────────────────────────────────────────
# Browser Automation
# ─────────────────────────────────────────────
playwright==1.45.0

# ─────────────────────────────────────────────
# Telegram Interface
# ─────────────────────────────────────────────
python-telegram-bot==21.4

# ─────────────────────────────────────────────
# Tool Execution / System
# ─────────────────────────────────────────────
psutil==6.0.0

# ─────────────────────────────────────────────
# Web API (if exposing server)
# ─────────────────────────────────────────────
fastapi==0.112.0
uvicorn[standard]==0.30.1

# ─────────────────────────────────────────────
# Observability
# ─────────────────────────────────────────────
prometheus-client==0.20.0

# ─────────────────────────────────────────────
# Testing
# ─────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
pydantic-settings==2.3.4
structlog==24.1.0